requirements.txt
asgiref==3.8.1
Django==4.2
django-debug-toolbar==4.2.0
django-filter==25.1
django-smart-selects==1.7.2
sqlparse==0.5.3
